introduced activerecord activerelation and query traits
