+++ framework/yii/redis/ActiveQuery.php
@@ -6,7 +6,8 @@
  */
 
 namespace yii\redis;
-use yii\base\NotSupportedException;
+
+use yii\ar\BaseActiveQuery;
 
 /**
  * ActiveQuery represents a query associated with an Active Record class.
@@ -43,79 +44,9 @@ use yii\base\NotSupportedException;
  * @author Carsten Brandt <mail@cebe.cc>
  * @since 2.0
  */
-class ActiveQuery extends \yii\base\Component
+class ActiveQuery extends Query
 {
-	/**
-	 * Sort ascending
-	 * @see orderBy
-	 */
-	const SORT_ASC = false;
-	/**
-	 * Sort descending
-	 * @see orderBy
-	 */
-	const SORT_DESC = true;
-
-	/**
-	 * @var string the name of the ActiveRecord class.
-	 */
-	public $modelClass;
-	/**
-	 * @var array list of relations that this query should be performed with
-	 */
-	public $with;
-	/**
-	 * @var string|callable $column the name of the column by which the query results should be indexed by.
-	 * This can also be a callable (e.g. anonymous function) that returns the index value based on the given
-	 * row or model data. For more details, see [[indexBy()]].
-	 */
-	public $indexBy;
-	/**
-	 * @var boolean whether to return each record as an array. If false (default), an object
-	 * of [[modelClass]] will be created to represent each record.
-	 */
-	public $asArray;
-	/**
-	 * @var array the query condition.
-	 * @see where()
-	 */
-	public $where;
-	/**
-	 * @var integer maximum number of records to be returned. If not set or less than 0, it means no limit.
-	 */
-	public $limit;
-	/**
-	 * @var integer zero-based offset from where the records are to be returned.
-	 * If not set, it means starting from the beginning.
-	 * If less than zero it means starting n elements from the end.
-	 */
-	public $offset;
-	/**
-	 * @var array how to sort the query results. This is used to construct the ORDER BY clause in a SQL statement.
-	 * The array keys are the columns to be sorted by, and the array values are the corresponding sort directions which
-	 * can be either [[ActiveQuery::SORT_ASC]] or [[ActiveQuery::SORT_DESC]]. The array may also contain [[Expression]] objects.
-	 * If that is the case, the expressions will be converted into strings without any change.
-	 */
-	public $orderBy;
-
-	/**
-	 * PHP magic method.
-	 * This method allows calling static method defined in [[modelClass]] via this query object.
-	 * It is mainly implemented for supporting the feature of scope.
-	 * @param string $name the method name to be called
-	 * @param array $params the parameters passed to the method
-	 * @return mixed the method return result
-	 */
-	public function __call($name, $params)
-	{
-		if (method_exists($this->modelClass, $name)) {
-			array_unshift($params, $this);
-			call_user_func_array(array($this->modelClass, $name), $params);
-			return $this;
-		} else {
-			return parent::__call($name, $params);
-		}
-	}
+	use BaseActiveQuery;
 
 	/**
 	 * Executes query and returns all results as an array.
@@ -191,8 +122,6 @@ class ActiveQuery extends \yii\base\Component
 
 	/**
 	 * Returns the number of records.
-	 * @param string $q the COUNT expression. Defaults to '*'.
-	 * Make sure you properly quote column names.
 	 * @return integer number of records
 	 */
 	public function count()
@@ -398,352 +327,5 @@ class ActiveQuery extends \yii\base\Component
 			}
 		}
 		return false;
-	}
-
-	// TODO: refactor. code below here is all duplicated from yii/db/ActiveQuery and yii/db/Query
-
-	/**
-	 * Sets the [[asArray]] property.
-	 * @param boolean $value whether to return the query results in terms of arrays instead of Active Records.
-	 * @return ActiveQuery the query object itself
-	 */
-	public function asArray($value = true)
-	{
-		$this->asArray = $value;
-		return $this;
-	}
-
-	/**
-	 * Sets the ORDER BY part of the query.
-	 * @param string|array $columns the columns (and the directions) to be ordered by.
-	 * Columns can be specified in either a string (e.g. "id ASC, name DESC") or an array
-	 * (e.g. `array('id' => Query::SORT_ASC, 'name' => Query::SORT_DESC)`).
-	 * The method will automatically quote the column names unless a column contains some parenthesis
-	 * (which means the column contains a DB expression).
-	 * @return ActiveQuery the query object itself
-	 * @see addOrderBy()
-	 */
-	public function orderBy($columns)
-	{
-		$this->orderBy = $this->normalizeOrderBy($columns);
-		return $this;
-	}
-
-	/**
-	 * Adds additional ORDER BY columns to the query.
-	 * @param string|array $columns the columns (and the directions) to be ordered by.
-	 * Columns can be specified in either a string (e.g. "id ASC, name DESC") or an array
-	 * (e.g. `array('id' => Query::SORT_ASC, 'name' => Query::SORT_DESC)`).
-	 * The method will automatically quote the column names unless a column contains some parenthesis
-	 * (which means the column contains a DB expression).
-	 * @return ActiveQuery the query object itself
-	 * @see orderBy()
-	 */
-	public function addOrderBy($columns)
-	{
-		$columns = $this->normalizeOrderBy($columns);
-		if ($this->orderBy === null) {
-			$this->orderBy = $columns;
-		} else {
-			$this->orderBy = array_merge($this->orderBy, $columns);
-		}
-		return $this;
-	}
-
-	protected function normalizeOrderBy($columns)
-	{
-		throw new NotSupportedException('orderBy is currently not supported');
-		if (is_array($columns)) {
-			return $columns;
-		} else {
-			$columns = preg_split('/\s*,\s*/', trim($columns), -1, PREG_SPLIT_NO_EMPTY);
-			$result = array();
-			foreach ($columns as $column) {
-				if (preg_match('/^(.*?)\s+(asc|desc)$/i', $column, $matches)) {
-					$result[$matches[1]] = strcasecmp($matches[2], 'desc') ? self::SORT_ASC : self::SORT_DESC;
-				} else {
-					$result[$column] = self::SORT_ASC;
-				}
-			}
-			return $result;
-		}
-	}
-
-	/**
-	 * Sets the LIMIT part of the query.
-	 * @param integer $limit the limit
-	 * @return ActiveQuery the query object itself
-	 */
-	public function limit($limit)
-	{
-		$this->limit = $limit;
-		return $this;
-	}
-
-	/**
-	 * Sets the OFFSET part of the query.
-	 * @param integer $offset the offset
-	 * @return ActiveQuery the query object itself
-	 */
-	public function offset($offset)
-	{
-		$this->offset = $offset;
-		return $this;
-	}
-
-	/**
-	 * Specifies the relations with which this query should be performed.
-	 *
-	 * The parameters to this method can be either one or multiple strings, or a single array
-	 * of relation names and the optional callbacks to customize the relations.
-	 *
-	 * The followings are some usage examples:
-	 *
-	 * ~~~
-	 * // find customers together with their orders and country
-	 * Customer::find()->with('orders', 'country')->all();
-	 * // find customers together with their country and orders of status 1
-	 * Customer::find()->with(array(
-	 *     'orders' => function($query) {
-	 *         $query->andWhere('status = 1');
-	 *     },
-	 *     'country',
-	 * ))->all();
-	 * ~~~
-	 *
-	 * @return ActiveQuery the query object itself
-	 */
-	public function with()
-	{
-		$this->with = func_get_args();
-		if (isset($this->with[0]) && is_array($this->with[0])) {
-			// the parameter is given as an array
-			$this->with = $this->with[0];
-		}
-		return $this;
-	}
-
-	/**
-	 * Sets the [[indexBy]] property.
-	 * @param string|callable $column the name of the column by which the query results should be indexed by.
-	 * This can also be a callable (e.g. anonymous function) that returns the index value based on the given
-	 * row or model data. The signature of the callable should be:
-	 *
-	 * ~~~
-	 * // $model is an AR instance when `asArray` is false,
-	 * // or an array of column values when `asArray` is true.
-	 * function ($model)
-	 * {
-	 *     // return the index value corresponding to $model
-	 * }
-	 * ~~~
-	 *
-	 * @return ActiveQuery the query object itself
-	 */
-	public function indexBy($column)
-	{
-		$this->indexBy = $column;
-		return $this;
-	}
-
-	/**
-	 * Sets the WHERE part of the query.
-	 *
-	 * The method requires a $condition parameter, and optionally a $params parameter
-	 * specifying the values to be bound to the query.
-	 *
-	 * The $condition parameter should be either a string (e.g. 'id=1') or an array.
-	 * If the latter, it must be in one of the following two formats:
-	 *
-	 * - hash format: `array('column1' => value1, 'column2' => value2, ...)`
-	 * - operator format: `array(operator, operand1, operand2, ...)`
-	 *
-	 * A condition in hash format represents the following SQL expression in general:
-	 * `column1=value1 AND column2=value2 AND ...`. In case when a value is an array,
-	 * an `IN` expression will be generated. And if a value is null, `IS NULL` will be used
-	 * in the generated expression. Below are some examples:
-	 *
-	 * - `array('type' => 1, 'status' => 2)` generates `(type = 1) AND (status = 2)`.
-	 * - `array('id' => array(1, 2, 3), 'status' => 2)` generates `(id IN (1, 2, 3)) AND (status = 2)`.
-	 * - `array('status' => null) generates `status IS NULL`.
-	 *
-	 * A condition in operator format generates the SQL expression according to the specified operator, which
-	 * can be one of the followings:
-	 *
-	 * - `and`: the operands should be concatenated together using `AND`. For example,
-	 * `array('and', 'id=1', 'id=2')` will generate `id=1 AND id=2`. If an operand is an array,
-	 * it will be converted into a string using the rules described here. For example,
-	 * `array('and', 'type=1', array('or', 'id=1', 'id=2'))` will generate `type=1 AND (id=1 OR id=2)`.
-	 * The method will NOT do any quoting or escaping.
-	 *
-	 * - `or`: similar to the `and` operator except that the operands are concatenated using `OR`.
-	 *
-	 * - `between`: operand 1 should be the column name, and operand 2 and 3 should be the
-	 * starting and ending values of the range that the column is in.
-	 * For example, `array('between', 'id', 1, 10)` will generate `id BETWEEN 1 AND 10`.
-	 *
-	 * - `not between`: similar to `between` except the `BETWEEN` is replaced with `NOT BETWEEN`
-	 * in the generated condition.
-	 *
-	 * - `in`: operand 1 should be a column or DB expression, and operand 2 be an array representing
-	 * the range of the values that the column or DB expression should be in. For example,
-	 * `array('in', 'id', array(1, 2, 3))` will generate `id IN (1, 2, 3)`.
-	 * The method will properly quote the column name and escape values in the range.
-	 *
-	 * - `not in`: similar to the `in` operator except that `IN` is replaced with `NOT IN` in the generated condition.
-	 *
-	 * - `like`: operand 1 should be a column or DB expression, and operand 2 be a string or an array representing
-	 * the values that the column or DB expression should be like.
-	 * For example, `array('like', 'name', '%tester%')` will generate `name LIKE '%tester%'`.
-	 * When the value range is given as an array, multiple `LIKE` predicates will be generated and concatenated
-	 * using `AND`. For example, `array('like', 'name', array('%test%', '%sample%'))` will generate
-	 * `name LIKE '%test%' AND name LIKE '%sample%'`.
-	 * The method will properly quote the column name and escape values in the range.
-	 *
-	 * - `or like`: similar to the `like` operator except that `OR` is used to concatenate the `LIKE`
-	 * predicates when operand 2 is an array.
-	 *
-	 * - `not like`: similar to the `like` operator except that `LIKE` is replaced with `NOT LIKE`
-	 * in the generated condition.
-	 *
-	 * - `or not like`: similar to the `not like` operator except that `OR` is used to concatenate
-	 * the `NOT LIKE` predicates.
-	 *
-	 * @param string|array $condition the conditions that should be put in the WHERE part.
-	 * @return ActiveQuery the query object itself
-	 * @see andWhere()
-	 * @see orWhere()
-	 */
-	public function where($condition)
-	{
-		$this->where = $condition;
-		return $this;
-	}
-
-	/**
-	 * Adds an additional WHERE condition to the existing one.
-	 * The new condition and the existing one will be joined using the 'AND' operator.
-	 * @param string|array $condition the new WHERE condition. Please refer to [[where()]]
-	 * on how to specify this parameter.
-	 * @return ActiveQuery the query object itself
-	 * @see where()
-	 * @see orWhere()
-	 */
-	public function andWhere($condition)
-	{
-		if ($this->where === null) {
-			$this->where = $condition;
-		} else {
-			$this->where = array('and', $this->where, $condition);
-		}
-		return $this;
-	}
-
-	/**
-	 * Adds an additional WHERE condition to the existing one.
-	 * The new condition and the existing one will be joined using the 'OR' operator.
-	 * @param string|array $condition the new WHERE condition. Please refer to [[where()]]
-	 * on how to specify this parameter.
-	 * @return ActiveQuery the query object itself
-	 * @see where()
-	 * @see andWhere()
-	 */
-	public function orWhere($condition)
-	{
-		if ($this->where === null) {
-			$this->where = $condition;
-		} else {
-			$this->where = array('or', $this->where, $condition);
-		}
-		return $this;
-	}
-
-	private function createModels($rows)
-	{
-		$models = array();
-		if ($this->asArray) {
-			if ($this->indexBy === null) {
-				return $rows;
-			}
-			foreach ($rows as $row) {
-				if (is_string($this->indexBy)) {
-					$key = $row[$this->indexBy];
-				} else {
-					$key = call_user_func($this->indexBy, $row);
-				}
-				$models[$key] = $row;
-			}
-		} else {
-			/** @var $class ActiveRecord */
-			$class = $this->modelClass;
-			if ($this->indexBy === null) {
-				foreach ($rows as $row) {
-					$models[] = $class::create($row);
-				}
-			} else {
-				foreach ($rows as $row) {
-					$model = $class::create($row);
-					if (is_string($this->indexBy)) {
-						$key = $model->{$this->indexBy};
-					} else {
-						$key = call_user_func($this->indexBy, $model);
-					}
-					$models[$key] = $model;
-				}
-			}
-		}
-		return $models;
-	}
-
-	private function populateRelations(&$models, $with)
-	{
-		$primaryModel = new $this->modelClass;
-		$relations = $this->normalizeRelations($primaryModel, $with);
-		foreach ($relations as $name => $relation) {
-			if ($relation->asArray === null) {
-				// inherit asArray from primary query
-				$relation->asArray = $this->asArray;
-			}
-			$relation->findWith($name, $models);
-		}
-	}
-
-	/**
-	 * @param ActiveRecord $model
-	 * @param array $with
-	 * @return ActiveRelation[]
-	 */
-	private function normalizeRelations($model, $with)
-	{
-		$relations = array();
-		foreach ($with as $name => $callback) {
-			if (is_integer($name)) {
-				$name = $callback;
-				$callback = null;
-			}
-			if (($pos = strpos($name, '.')) !== false) {
-				// with sub-relations
-				$childName = substr($name, $pos + 1);
-				$name = substr($name, 0, $pos);
-			} else {
-				$childName = null;
-			}
-
-			$t = strtolower($name);
-			if (!isset($relations[$t])) {
-				$relation = $model->getRelation($name);
-				$relation->primaryModel = null;
-				$relations[$t] = $relation;
-			} else {
-				$relation = $relations[$t];
-			}
-
-			if (isset($childName)) {
-				$relation->with[$childName] = $callback;
-			} elseif ($callback !== null) {
-				call_user_func($callback, $relation);
-			}
-		}
-		return $relations;
 	}
 }
